use single DAG for database setup
